Minio.py/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
marshmallow
python-dotenv==0.19.2
Werkzeug==2.0.2
flask-marshmallow
flask_swagger_ui==3.36.0
requests==2.26.0
jsonschema==4.3.0
minio==7.1.2
kafka-python==2.0.2
Pillow==8.4.0
pymongo==4.0.1